// src/shared/api/household-calculator-client/models/Household.ts
export interface Household {
    /**
     * Where the household is located
     * @type {string}
     * @memberof Household
     */
    location?: HouseholdLocationEnum;
    /**
     * Number of occupants
     * @type {number}
     * @memberof Household
     */
    occupancy?: number;
    /**
     * The main method of space heating
     * @type {string}
     * @memberof Household
     */
    spaceHeating?: HouseholdSpaceHeatingEnum;
    /**
     * The method of water heating
     * @type {string}
     * @memberof Household
     */
    waterHeating?: HouseholdWaterHeatingEnum;
    /**
     * The main energy source for cooking
     * @type {string}
     * @memberof Household
     */
    cooktop?: HouseholdCooktopEnum;
    /**
     * 
     * @type {Array<Vehicle>}
     * @memberof Household
     */
    vehicles?: Array<Vehicle>;
    /**
     * 
     * @type {Solar}
     * @memberof Household
     */
    solar?: Solar;
    /**
     * 
     * @type {Battery}
     * @memberof Household
     */
    battery?: Battery;
}


/**
 * @export
 */
export const HouseholdLocationEnum = {
    Northland: 'NORTHLAND',
    AucklandNorth: 'AUCKLAND_NORTH',
    AucklandCentral: 'AUCKLAND_CENTRAL',
    AucklandEast: 'AUCKLAND_EAST',
    AucklandWest: 'AUCKLAND_WEST',
    AucklandSouth: 'AUCKLAND_SOUTH',
    Waikato: 'WAIKATO',
    BayOfPlenty: 'BAY_OF_PLENTY',
    Gisborne: 'GISBORNE',
    HawkesBay: 'HAWKES_BAY',
    Taranaki: 'TARANAKI',
    ManawatuWanganui: 'MANAWATU_WANGANUI',
    Wellington: 'WELLINGTON',
    Tasman: 'TASMAN',
    Nelson: 'NELSON',
    Marlborough: 'MARLBOROUGH',
    WestCoast: 'WEST_COAST',
    Canterbury: 'CANTERBURY',
    Otago: 'OTAGO',
    Southland: 'SOUTHLAND',
    StewartIsland: 'STEWART_ISLAND',
    ChathamIslands: 'CHATHAM_ISLANDS',
    GreatBarrierIsland: 'GREAT_BARRIER_ISLAND',
    Overseas: 'OVERSEAS',
    Other: 'OTHER'
} as const;
export type HouseholdLocationEnum = typeof HouseholdLocationEnum[keyof typeof HouseholdLocationEnum];

/**
 * @export
 */
export const HouseholdSpaceHeatingEnum = {
    Wood: 'WOOD',
    Gas: 'GAS',
    Lpg: 'LPG',
    ElectricResistance: 'ELECTRIC_RESISTANCE',
    ElectricHeatPump: 'ELECTRIC_HEAT_PUMP',
    DontKnow: 'DONT_KNOW'
} as const;
export type HouseholdSpaceHeatingEnum = typeof HouseholdSpaceHeatingEnum[keyof typeof HouseholdSpaceHeatingEnum];

/**
 * @export
 */
export const HouseholdWaterHeatingEnum = {
    Gas: 'GAS',
    Lpg: 'LPG',
    ElectricResistance: 'ELECTRIC_RESISTANCE',
    ElectricHeatPump: 'ELECTRIC_HEAT_PUMP',
    Solar: 'SOLAR',
    DontKnow: 'DONT_KNOW'
} as const;
export type HouseholdWaterHeatingEnum = typeof HouseholdWaterHeatingEnum[keyof typeof HouseholdWaterHeatingEnum];

/**
 * @export
 */
export const HouseholdCooktopEnum = {
    Gas: 'GAS',
    Lpg: 'LPG',
    ElectricResistance: 'ELECTRIC_RESISTANCE',
    ElectricHeatPump: 'ELECTRIC_HEAT_PUMP',
    DontKnow: 'DONT_KNOW'
} as const;
export type HouseholdCooktopEnum = typeof HouseholdCooktopEnum[keyof typeof HouseholdCooktopEnum];